backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
pillow==10.4.0
numpy==1.26.4
tensorflow==2.16.2
opencv-python==4.10.0.84
pydantic==2.9.2
python-dotenv==1.0.1
